chore: move logger from pkg to internal
