+++ cli/internal/providers/datacatalog/provider.go
@@ -5,26 +5,26 @@ import (
 	"strings"
 
 	"github.com/rudderlabs/rudder-iac/api/client/catalog"
+	"github.com/rudderlabs/rudder-iac/cli/internal/logger"
 	"github.com/rudderlabs/rudder-iac/cli/internal/project/specs"
 	pstate "github.com/rudderlabs/rudder-iac/cli/internal/providers/datacatalog/state"
 	"github.com/rudderlabs/rudder-iac/cli/internal/syncer/resources"
 	"github.com/rudderlabs/rudder-iac/cli/pkg/localcatalog"
-	"github.com/rudderlabs/rudder-iac/cli/pkg/logger"
 	"github.com/rudderlabs/rudder-iac/cli/pkg/validate"
 )
 
 var log = logger.New("datacatalogprovider")
 
 type Provider struct {
-	client catalog.DataCatalog
-	dc     *localcatalog.DataCatalog
+	client        catalog.DataCatalog
+	dc            *localcatalog.DataCatalog
 	providerStore map[string]resourceProvider
 }
 
 func New(client catalog.DataCatalog) *Provider {
 	return &Provider{
-		client:       client,
-		dc:           localcatalog.New(),
+		client: client,
+		dc:     localcatalog.New(),
 		providerStore: map[string]resourceProvider{
 			PropertyResourceType:     NewPropertyProvider(client),
 			EventResourceType:        NewEventProvider(client),
